fix: avoid deleting non-default namespace

+++ pkg/jobs/collector.go
@@ -14,7 +14,8 @@ import (
 )
 
 const (
-	NodeCollectorName = "node-collector"
+	NodeCollectorName             = "node-collector"
+	nodeCollectorDefaultNamespace = "trivy-temp"
 
 	// job headers
 	TrivyCollectorName = "trivy.collector.name"
@@ -362,5 +363,8 @@ func (jb *jobCollector) getTrivyNamespace(ctx context.Context) (*corev1.Namespac
 }
 
 func (jb *jobCollector) Cleanup(ctx context.Context) {
+	if jb.namespace != nodeCollectorDefaultNamespace {
+		return
+	}
 	jb.deleteTrivyNamespace(ctx)
 }
